Move all Page classes to pages.py
