+++ cli.py
@@ -69,6 +69,7 @@ def cli_entry_point(scrape, resume, verify, compile_active, dm, tf, rn, registra
             click.echo('File Not Found for the task.')
         return
 
+
     return
